ApiReport. Move data-classes to shared namespace
