+++ src/back/External/ApiReport/Actions/CsvReportProcessor.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace KeepersTeam\Webtlo\External\ApiReport\Actions;
 
 use Closure;
-use KeepersTeam\Webtlo\External\ApiReport\V1\KeeperTopics;
-use KeepersTeam\Webtlo\External\ApiReport\V1\KeptTopic;
 use KeepersTeam\Webtlo\External\Contract\ReportProcessorInterface;
+use KeepersTeam\Webtlo\External\Data\KeeperTopics;
+use KeepersTeam\Webtlo\External\Data\KeptTopic;
 use League\Csv\Reader;
 
 final class CsvReportProcessor implements ReportProcessorInterface
