+++ src/back/Storage/Clone/KeepersSeeders.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace KeepersTeam\Webtlo\Storage\Clone;
 
 use KeepersTeam\Webtlo\Data\Keeper;
-use KeepersTeam\Webtlo\External\ApiReport\V1\KeptTopic;
+use KeepersTeam\Webtlo\External\Data\KeptTopic;
 use KeepersTeam\Webtlo\Storage\CloneTable;
 use Psr\Log\LoggerInterface;
